index-dump: fix argument naming, resolve Swift 6 build errors, and cleanup descriptions
