+++ Sources/index-dump/main.swift
@@ -27,10 +27,10 @@ struct IndexDump: AsyncParsableCommand {
   @Argument(help: "Name of the unit/record or a direct path to the file")
   var nameOrPath: String
 
-  @Option(help: "Path to libIndexStore. Inferred if omitted.")
+  @Option(name: .customLong("lib-index-store"), help: "Path to libIndexStore. Inferred if omitted.")
   var libIndexStore: String?
 
-  @Option(help: "Path to the index store directory. Inferred if omitted.")
+  @Option(name: .customLong("index-store"), help: "Path to the index store directory. Inferred if omitted.")
   var indexStore: String?
 
   @Option(help: "Explicitly set mode (unit/record).")
@@ -89,6 +89,67 @@ struct IndexDump: AsyncParsableCommand {
       throw ValidationError(
         "Could not find 'swift' to infer toolchain path. Please specify --lib-index-store explicitly."
       )
+    }
+    return libURL
+  }
+}
+
+// MARK: - LibIndexStoreProvider
+
+/// Provides utilities to discover and locate libIndexStore in the system.
+private enum LibIndexStoreProvider {
+  /// Find a tool using xcrun/which/where (copied logic from TibsToolchain.findTool)
+  static func findTool(name: String) -> URL? {
+    #if os(macOS)
+    let cmd = ["/usr/bin/xcrun", "--find", name]
+    #elseif os(Windows)
+    var buf = [WCHAR](repeating: 0, count: Int(MAX_PATH))
+    GetWindowsDirectoryW(&buf, UINT(MAX_PATH))
+    var wherePath = String(decodingCString: &buf, as: UTF16.self)
+    wherePath = (wherePath as NSString).appendingPathComponent("system32")
+    wherePath = (wherePath as NSString).appendingPathComponent("where.exe")
+    let cmd = [wherePath, name]
+    #else
+    let cmd = ["/usr/bin/which", name]
+    #endif
+
+    let process = Process()
+    process.executableURL = URL(fileURLWithPath: cmd[0])
+    process.arguments = Array(cmd.dropFirst())
+    let pipe = Pipe()
+    process.standardOutput = pipe
+
+    try? process.run()
+    process.waitUntilExit()
+    guard process.terminationStatus == 0 else { return nil }
+
+    let data = pipe.fileHandleForReading.readDataToEndOfFile()
+    var path = String(decoding: data, as: UTF8.self)
+    #if os(Windows)
+    path = String((path.split { $0.isNewline })[0])
+    #endif
+    return URL(fileURLWithPath: path.trimmingCharacters(in: .whitespacesAndNewlines))
+  }
+
+  /// Infer libIndexStore dylib path from the default toolchain
+  static func inferLibPath() -> URL? {
+    guard let swiftURL = findTool(name: "swift") else {
+      return nil
+    }
+
+    let toolchainURL = swiftURL.deletingLastPathComponent().deletingLastPathComponent()
+
+    #if os(macOS)
+    let libName = "libIndexStore.dylib"
+    #elseif os(Windows)
+    let libName = "IndexStore.dll"
+    #else
+    let libName = "libIndexStore.so"
+    #endif
+
+    let libURL = toolchainURL.appending(components: "lib", libName)
+    guard FileManager.default.fileExists(atPath: libURL.path) else {
+      return nil
     }
     return libURL
   }
